renamed splitter base classes file
