+++ macrosynergy/learning/splitters/walk_forward_splitters.py
@@ -7,7 +7,7 @@ import pandas as pd
 
 from typing import List
 
-from macrosynergy.learning.splitters.base_panel_split import WalkForwardPanelSplit
+from macrosynergy.learning.splitters.base_splitters import WalkForwardPanelSplit
 
 class ExpandingIncrementPanelSplit(WalkForwardPanelSplit):
     """
